+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ### 3.5.2 Release (20.08.2018)
 * Fixed NullPointerException for users who where no longer online
+* Now using default values when they not exist in config to avoid NullPointerException from ConfigPreferences
 
 ### 3.5.1 Release (17.08.2018)
 * Fixed NumberFormatException for language.yml migrator - this error is very rare to occur in normal environment but it was reported so I fixed it